qrr/requirements.txt
qrcode
Pillow 